+++ SIM-VICUS/src/core3D/Vic3DOpaqueGeometryObject.h
@@ -66,7 +66,7 @@ public:
 	/*! Color buffer in CPU memory, holds colors of all vertices (same size as m_vertexBufferData). */
 	std::vector<ColorRGBA>		m_colorBufferData;
 	/*! Index buffer on CPU memory. */
-	std::vector<GLshort>		m_indexBufferData;
+	std::vector<GLuint>			m_indexBufferData;
 
 	/*! Maps unique surface/node ID to vertex start index in m_vertexBufferData. */
 	std::map<unsigned int, unsigned int>	m_vertexStartMap;
